commit end day

+++ ap/src/main/java/com/brief/TDD/repository/ClientRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.brief.TDD.entity.Clients;
 
 @Repository
-public interface ClientRepository extends JpaRepository<Clients, Long> {
+public interface ClientRepository extends JpaRepository<Clients, int > {
 	   // No additional methods needed for this example
 
 }
